+++ src/main/java/co/invest72/exchange_rate/domain/ExchangeRateProvider.java
@@ -1,7 +1,10 @@
-package co.invest72.money.domain;
+package co.invest72.exchange_rate.domain;
 
 import java.math.BigDecimal;
 import java.util.Optional;
+
+import co.invest72.money.domain.Currency;
+import co.invest72.money.domain.Pair;
 
 public interface ExchangeRateProvider {
 	Optional<BigDecimal> getRate(Currency from, Currency to);
